# docs/swagger.yaml
basePath: /ih/mobile_central_2_0_0
definitions:
...
  dto.LogisticRequestDto:
    properties:
      companionDob:
        type: string
      companionDocNumber:
        type: string
      companionDocType:
        type: string
      companionName:
        type: string
      dateCreate:
        type: string
      dateUpdate:
        type: string
      flightAirlineName:
        type: string
      flightArrivalDate:
        type: string
      flightArrivalTime:
        type: string
      flightNumber:
        type: string
      primaryDoctor:
        type: string
      relationshipToRequester:
        type: string
      requestedPickupDate:
        type: string
      requestedPickupTime:
        type: string
      requesterDob:
        type: string
      requesterDocNumber:
        type: string
      requesterDocType:
        type: string
      requesterEmail:
        type: string
      requesterName:
        type: string
      requesterNationality:
        type: string
      requesterPrn:
        type: string
      visitWithCompanion:
        type: string
    required:
    - flightAirlineName
    - flightArrivalDate
    - flightArrivalTime
    - flightNumber
    - primaryDoctor
    - requestedPickupDate
    - requestedPickupTime
    - requesterDocNumber
    - requesterDocType
    - requesterEmail
    - requesterName
    - requesterNationality
    - visitWithCompanion
    type: object
  dto.LogisticRequestStatusDto:
    properties:
      requestNumber:
        type: string
      status:
        type: string
    required:
    - requestNumber
    - status
    type: object
  dto.LogisticSetupDto:
    properties:
      logisticSetupValue:
        type: string
    required:
    - logisticSetupValue
    type: object
  dto.LogisticSlotDto:
    properties:
      dayOfWeek:
        type: string
      displaySequence:
        type: integer
      maxSlots:
        type: integer
      pickUpTime:
        type: string
    required:
    - dayOfWeek
    - maxSlots
    - pickUpTime
    type: object
  dto.LogisticSlotMobileDto:
    properties:
      flightArrivalDate:
        type: string
      flightArrivalTime:
        type: string
      withCompanion:
        type: boolean
    required:
    - flightArrivalDate
    - flightArrivalTime
    - withCompanion
    type: object
  dto.LogisticSlotsDto:
    properties:
      logisticSlots:
        items:
          $ref: '#/definitions/dto.LogisticSlotDto'
        type: array
    type: object
...
  dto.SearchKeyword4Dto:
    properties:
      keyword:
        type: string
      keyword2:
        type: string
      keyword3:
        type: string
      keyword4:
        type: string
    type: object
...
  logistic.LogisticRequest:
    properties:
      adminDateUpdate:
        type: string
      adminDateUpdateExcel:
        type: string
      adminUpdate:
        type: string
      companionDob:
        type: string
      companionDocNumber:
        type: string
      companionDocType:
        type: string
      companionName:
        type: string
      dateCreate:
        type: string
      dateCreateExcel:
        type: string
      displayCancelBtn:
        type: boolean
      flightAirlineName:
        type: string
      flightArrivalDate:
        type: string
      flightArrivalDateExcel:
        type: string
      flightArrivalTime:
        type: string
      flightNumber:
        type: string
      logistic_request_id:
        type: integer
      logisticRequestNumber:
        type: string
      logisticRequestStatus:
        type: string
      primaryDoctor:
        type: string
      primaryDoctorName:
        type: string
      relationshipToRequester:
        type: string
      requestedPickupDate:
        type: string
      requestedPickupDateExcel:
        type: string
      requestedPickupDay:
        type: string
      requestedPickupTime:
        type: string
      requesterDob:
        type: string
      requesterDocNumber:
        type: string
      requesterDocType:
        type: string
      requesterEmail:
        type: string
      requesterName:
        type: string
      requesterNationality:
        type: string
      requesterPrn:
        type: string
      userDateUpdate:
        type: string
      userDateUpdateExcel:
        type: string
      userUpdate:
        type: string
      visitWithCompanion:
        type: string
    type: object
  logistic.LogisticSetup:
    properties:
      logistic_setup_id:
        type: string
      logisticSetupCode:
        type: string
      logisticSetupValue:
        type: string
    type: object
  logistic.LogisticSlot:
    properties:
      availableSlots:
        type: integer
      dayOfWeek:
        type: string
      displaySequence:
        type: integer
      logistic_slot_id:
        type: integer
      maxSlots:
        type: integer
      pickUpDate:
        type: string
      pickUpTime:
        type: string
    type: object
...
info:
  contact: {}
  description: Vesalius-m Backend API Docs (IHP-UAT Version).
  title: (IHP-UAT) Vesalius-m Backend API
  version: "1.0"
paths:
...
  /logistic/request:
    post:
      parameters:
      - description: LogisticRequestDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/{requestId}:
    get:
      parameters:
      - description: requestId
        in: path
        name: requestId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/logistic.LogisticRequest'
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/all:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticRequest'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
    post:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      - description: Keyword
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.SearchKeyword4Dto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticRequest'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/all/mobile:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticRequest'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/export/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticRequest'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/export/search:
    post:
      parameters:
      - description: Search
        in: body
        name: keyword
        schema:
          $ref: '#/definitions/dto.SearchKeyword4Dto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticRequest'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/status:
    post:
      parameters:
      - description: LogisticRequestStatusDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticRequestStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/request/status/webadmin:
    post:
      parameters:
      - description: LogisticRequestStatusDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticRequestStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/setup:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/logistic.LogisticSetup'
      security:
      - BearerAuth: []
      tags:
      - Logistic
    post:
      parameters:
      - description: LogisticSetupDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticSetupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/setup/{logisticSetupId}:
    put:
      parameters:
      - description: logisticSetupId
        in: path
        name: logisticSetupId
        required: true
        type: string
      - description: LogisticSetupDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticSetupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/slot:
    post:
      parameters:
      - description: LogisticSlotsDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticSlotsDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/slot/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticSlot'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
  /logistic/slot/all/mobile:
    post:
      parameters:
      - description: LogisticSlotMobileDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LogisticSlotMobileDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logistic.LogisticSlot'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Logistic
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"